Update SCCM ToolKit.ps1
Updated SCCMToolKit .

Added check WMI button that checks one aspect of WMI. And another of SCCM locally.
If they fail or pass it passes through a scripted voting system. Automatically deciding how to resolve the issue.
If the SCCM client needs reinstalled I do not allow the script to do this automatically and you will have to do it manually.

Also added a check SCCM button.
Checks the CM client to see if it is in debug mode. Also checks the CcmSQLCE.log file. If so warns and starts the SDF file cleanup and reinstall of the CM client.

Then stops the service and deletes the *.sdf files after verifying the service is stopped.

I do not delete the client automatically in case you are remotely connected to the PC.

This is due to the client connecting to the SCCM server and auto populating the file from the server on the next connection.

Computer Info window now populates my very own inventory gathering script, modified to fit in a smaller window, and is another good way to test other aspects of WMI failures.
                Given the different ways I pull WMI info for the Computer Info it really does cover the gambit of issues we have seen.
Added CCMsetup log button to open CM trace and the log file automatically.
Added CCMrepair log button to open CM trace and the log file automatically.
Added Repair SCCM Button to do just that.
Added CM Trace button to open CM trace.